[storage] One sync::Database impl for the compact dbs
The keyless and immutable compact dbs carried identical `sync::Database` impls (keyless's `V: Codec` bound was redundant with `Operation: Read`). One impl on the shared `compact::db::Db` replaces both, and the `from_sync_result` helper the two impls shared is inlined into it, its only remaining caller.
